Fix error prone code

+++ src/listening/phd/handshake.c
@@ -57,7 +57,7 @@ void phd_send_legacy_slp(ltg_client_t* client) {
 
 	size_t length = 0;
 	with_lock (&sky_main.listener.online.lock) {
-		length = sprintf(legacy_slp, "\xa7\x31%c127%cMotor MC " __MC_VER__ "%c%s%c%u%c%llu", '\0', '\0', '\0', motd, '\0', sky_main.listener.online.list.length, '\0', (size_t) sky_main.listener.online.max);
+		length = sprintf(legacy_slp, "\xa7\x31%c127%cMotor MC " __MC_VER__ "%c%s%c%u%c%u", '\0', '\0', '\0', motd, '\0', sky_main.listener.online.list.length, '\0', sky_main.listener.online.max);
 	}
 	pck_write_int16(packet, length);
 	for (size_t i = 0; i < length; ++i) {
